Run conformance tests as part of runtime tests
