--- tests/runtime-tests/tests/key-value-no-permission/error.txt
@@ -1,1 +1,0 @@
-Error::AccessDenied
