--- tests/runtime-tests/tests/key-value/spin.toml
@@ -1,14 +1,0 @@
-spin_manifest_version = 2
-
-[application]
-name = "key-value"
-authors = ["Fermyon Engineering <engineering@fermyon.com>"]
-version = "0.1.0"
-
-[[trigger.http]]
-route = "/"
-component = "test"
-
-[component.test]
-source = "%{source=key-value}"
-key_value_stores = ["default"]
