--- tests/runtime-tests/tests/outbound-mqtt-variable-permission/services
@@ -1,1 +1,0 @@
-mqtt
