--- tests/runtime-tests/tests/outbound-mqtt/services
@@ -1,1 +1,0 @@
-mqtt
