--- tests/runtime-tests/tests/outbound-redis-no-permission/error.txt
@@ -1,1 +1,0 @@
-Error::InvalidAddress
