--- tests/runtime-tests/tests/outbound-redis-no-permission/spin.toml
@@ -1,14 +1,0 @@
-spin_manifest_version = 2
-
-[application]
-name = "outbound-redis"
-authors = ["Fermyon Engineering <engineering@fermyon.com>"]
-version = "0.1.0"
-
-[[trigger.http]]
-route = "/"
-component = "test"
-
-[component.test]
-source = "%{source=outbound-redis}"
-environment = { REDIS_ADDRESS = "redis://localhost:6379" }
