--- tests/runtime-tests/tests/outbound-redis-variable-permission/services
@@ -1,1 +1,0 @@
-redis
