--- tests/runtime-tests/tests/outbound-redis-variable-permission/spin.toml
@@ -1,18 +1,0 @@
-spin_manifest_version = 2
-
-[application]
-name = "outbound-redis"
-authors = ["Fermyon Engineering <engineering@fermyon.com>"]
-version = "0.1.0"
-
-[variables]
-redis_host = { default = "localhost" }
-
-[[trigger.http]]
-route = "/"
-component = "test"
-
-[component.test]
-source = "%{source=outbound-redis}"
-environment = { REDIS_ADDRESS = "redis://localhost:%{port=6379}" }
-allowed_outbound_hosts = ["redis://{{ redis_host }}:%{port=6379}"]
