--- tests/runtime-tests/tests/outbound-redis/services
@@ -1,1 +1,0 @@
-redis
