--- tests/runtime-tests/tests/sqlite-no-permission/error.txt
@@ -1,1 +1,0 @@
-Error::AccessDenied
