--- tests/runtime-tests/tests/tcp-sockets-ip-range-variable-permission/services
@@ -1,1 +1,0 @@
-tcp-echo
