--- tests/runtime-tests/tests/tcp-sockets-ip-range-variable-permission/spin.toml
@@ -1,19 +1,0 @@
-spin_manifest_version = 2
-
-[application]
-name = "tcp-sockets"
-authors = ["Fermyon Engineering <engineering@fermyon.com>"]
-version = "0.1.0"
-
-[variables]
-addr_prefix = { default = "127.0.0.0" }
-prefix_len = { default = "24" }
-
-[[trigger.http]]
-route = "/"
-component = "test"
-
-[component.test]
-source = "%{source=tcp-sockets}"
-environment = { ADDRESS = "127.0.0.1:%{port=7}" }
-allowed_outbound_hosts = ["*://{{ addr_prefix }}/{{ prefix_len }}:%{port=7}"]
