--- tests/runtime-tests/tests/tcp-sockets-ip-range/services
@@ -1,1 +1,0 @@
-tcp-echo
