--- tests/runtime-tests/tests/tcp-sockets-no-ip-permission/error.txt
@@ -1,1 +1,0 @@
-access-denied
