--- tests/runtime-tests/tests/tcp-sockets-no-ip-permission/spin.toml
@@ -1,16 +1,0 @@
-spin_manifest_version = 2
-
-[application]
-name = "tcp-sockets"
-authors = ["Fermyon Engineering <engineering@fermyon.com>"]
-version = "0.1.0"
-
-[[trigger.http]]
-route = "/"
-component = "test"
-
-[component.test]
-source = "%{source=tcp-sockets}"
-environment = { ADDRESS = "127.0.0.1:6001" }
-# Component expects 127.0.0.1 but we only allow 127.0.0.2
-allowed_outbound_hosts = ["*://127.0.0.2:6001"]
