--- tests/runtime-tests/tests/tcp-sockets-no-port-permission/error.txt
@@ -1,1 +1,0 @@
-access-denied
