--- tests/runtime-tests/tests/tcp-sockets/services
@@ -1,1 +1,0 @@
-tcp-echo
