--- tests/runtime-tests/tests/variables/spin.toml
@@ -1,16 +1,0 @@
-spin_manifest_version = 2
-
-[application]
-name = "variables"
-authors = ["Fermyon Engineering <engineering@fermyon.com>"]
-version = "0.1.0"
-
-[[trigger.http]]
-route = "/"
-component = "test"
-
-[component.test]
-source = "%{source=variables}"
-
-[component.test.variables]
-variable = "value"
